+++ src/main/java/com/codelad/authservice/services/JWTService.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.UUID;
 
 public interface JWTService {
-    public Integer extractUid(String token);
+    public Long extractUid(String token);
     public String generateToken(UserDetails userDetails);
     public boolean isTokenValid(String token, UserDetails userDetails);
 }
